fix(mobile): give the switcher's rows the desktop leading icon itself
The sidebar row's leading icon and its slot were module-private to
`epic-sidebar-chat-tree`, so the mobile switcher could not reach them and grew
a parallel `SwitcherAgentIcon` instead. Two components painting one row's
single status surface is a difference a user sees, and the mobile one was also
substantively poorer: it had no rollup, so a collapsed parent on the phone said
nothing about a failure hidden inside its subtree.

Extract them the way #1406 extracted the row menu. `chat-row-leading-icon`
holds the components; `chat-row-status-ladder` holds the ranks, the classifier
and the descendant walk, because the panel body classifies rows without
rendering an icon and a module that exports both loses fast refresh.

Desktop is a pure move - the same glyphs, slot, test ids and subscriptions, at
the same call sites. The switcher now renders that component with its own
`hasChildren`/`expanded`, so a collapsed row there resolves the rollup too. Its
status test ids become the sidebar's, since the ids belong to the component.
